+++ PolygonDrawer/Renderers/GdiRenderer.cs
@@ -18,14 +18,14 @@ namespace PolygonDrawer.Renderers
             _pointBrush = brush;
         }
 
-        public void DrawLine(int x1, int y1, int x2, int y2)
+        public void DrawLine(float x1, float y1, float x2, float y2)
         {
             _graphics?.DrawLine(Pens.Black, x1, y1, x2, y2);
         }
 
-        public void DrawPoint(int x, int y)
+        public void DrawPoint(float x, float y)
         {
-            if(_pointBrush is null)
+            if (_pointBrush is null)
             {
                 return;
             }
@@ -33,12 +33,27 @@ namespace PolygonDrawer.Renderers
             _graphics?.FillEllipse(_pointBrush, x - 3, y - 3, 6, 6);
         }
 
-        public void DrawBezierCurve(int x1, int y1, int x2, int y2, int cp1x, int cp1y, int cp2x, int cp2y)
+        public void DrawBezierCurve(
+            float x1,
+            float y1,
+            float x2,
+            float y2,
+            float cp1x,
+            float cp1y,
+            float cp2x,
+            float cp2y)
         {
             _graphics?.DrawBezier(Pens.Black, x1, y1, cp1x, cp1y, cp2x, cp2y, x2, y2);
         }
 
-        public void DrawCircle(int middlex, int middley, float radius, int xfrom, int yfrom, int xto, int yto)
+        public void DrawCircle(
+            float middlex,
+            float middley,
+            float radius,
+            float xfrom,
+            float yfrom,
+            float xto,
+            float yto)
         {
             var rect = new RectangleF(
                 middlex - radius,
